requirements.txt
langchain
python-dotenv
langchain-community
faiss-cpu
langchain-google-genai
google_generativeai
PyPDF2
streamlit